Refactor ht backends incl. py3 support
The error reportring didn't help much
as most error messages where caught and overwritten.
Errors that weren't caught resulted in real server error.
Additinally `status_check` returned a status code in addition
to raising errors.

`status_check` only raies errors now or returns `None` if everything
is already. Similar to how `ValidationErrors` are handle in Django.
You may raise a `HealthCheckError` at any time and it will be properly
caught and displayed in the health check view.

Other, or unexpected Exceptions are still caught and displayed, for
the health check view to never be down.

Additinally logging was added to get a better insight including a
stack trace to where the Exception was raised.

This change may only work in python 3.

Closes #2