Add a config setting for preferring JSON to HTML as the default output format.
As in the existing code, this doesn't try to fully parse the Accept header, but just uses simple substring matching within the header. As such, it doesn't take account of any q-values. If both text/html and application/json are present, it ignores the request's preferences and uses the configured preferred format.