Allow more valid q-values.
'[\d3]' is the range of digits or the digit 3, so we're only allowing a single decimal digit. RFC 2616 allows for up to three digits but no more after a zero, or up to three zeroes after a one.

The new pattern is cumbersome, but should allow any decimal from 0 to 1 (inclusive), with up to three digits of precision.