Allow more valid accept headers (#221)
* Allow any amount of optional whitespace on either side of the semi-colon for parameters.

* Allow other parameters after the optional q-value.

RFC 2616 says that the q comes first, which spares us needing to fully parse the list here.

The list of characters comes from RFC 7230, section 3.2.6 ('visible US-ASCII characters, except delimiters').

* Allow more valid q-values.

'[\d3]' is the range of digits or the digit 3, so we're only allowing a single decimal digit. RFC 2616 allows for up to three digits but no more after a zero, or up to three zeroes after a one.

The new pattern is cumbersome, but should allow any decimal from 0 to 1 (inclusive), with up to three digits of precision.