Improve GitHub Actions PyPI release trigger (refs #250)
Hey @codingjoe!

Thanks for actively maintaining this and plenty of other packages, you are the champion! :)
I was replacing Travis CI with GitHub actions in one of my packages (https://github.com/amureki/statuscheck) and was inspired by your code here in #250.

However, I noticed a couple of things that were not correct from my perspective:

- release job was triggered three times (fixed by specifying exact action type "released")
- TWINE_USERNAME variable was coming from `__token__` - not sure I follow, how this works, but I changed it to `secrets.TWINE_USERNAME`
- (other) I put a couple of naming suggestions, feel free to disregard