Move the description of `HEALTHCHECK_CACHE_KEY` to the correct section
It's a setting for the cache backend, but it's documented under "Celery Health
Check". Create a "cache" section, and move the documentation there.

It's also redundantly documented in `contrib.rst`, despite not being a contrib
module. Remove it from there.

Fix the typo in the default value while I'm at it.